mozio/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-geojson==3.2.0
djangorestframework==3.12.4
djangorestframework-gis==0.17
jsonfield==3.1.0
psycopg2==2.9.1
pytz==2021.1
Shapely==1.7.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
